[MONDRIAN-2529] ChooseByVolume=true can fail due to integer overflow

+++ mondrian/src/it/java/mondrian/test/BasicQueryTest.java
@@ -7913,7 +7913,7 @@ public class BasicQueryTest extends FoodMartTestCase {
                 statisticsProviders.get(1) instanceof SqlStatisticsProvider);
 
             for (StatisticsProvider statisticsProvider : statisticsProviders) {
-                int rowCount =
+                long rowCount =
                     statisticsProvider.getTableCardinality(
                         dialect,
                         testContext.getConnection().getDataSource(),
@@ -7932,7 +7932,7 @@ public class BasicQueryTest extends FoodMartTestCase {
                         rowCount > 10000 && rowCount < 15000);
                 }
 
-                int valueCount =
+                long valueCount =
                     statisticsProvider.getColumnCardinality(
                         dialect,
                         testContext.getConnection().getDataSource(),
